mission: move user_can_command_asset to helpers
Pure refactor. The search app needs this check for its abandon endpoint,
and importing it from a views module to do so would be the wrong seam.
